refactor(archon): extract inference bridge backends into sglang/vllm submodules
Break monolithic data_proxy/backend.py into properly-separated
submodules with a protocol-based backend abstraction, preparing
the inference_service package structure for weight update integration.

Key changes:
- Add InfBridgeBackend protocol in inference_service/backend.py
- Extract SGLangBridgeBackend into sglang/bridge.py
- Extract VLLMBridgeBackend into vllm/bridge.py
- Move inf_bridge.py from data_proxy/ to inference_service/
- Create sglang/ and vllm/ as proper Python packages
- Update all consumer imports (app.py, 9 test files)
